+++ activesupport/test/core_ext/object/blank_test.rb
@@ -33,4 +33,10 @@ class BlankTest < ActiveSupport::TestCase
     BLANK.each { |v| assert_equal nil, v.presence, "#{v.inspect}.presence should return nil" }
     NOT.each   { |v| assert_equal v,   v.presence, "#{v.inspect}.presence should return self" }
   end
+
+  def test_presence_or
+    alternative = 'ruby <3'
+    BLANK.each { |v| assert_equal alternative, v.presence_or(alternative), "#{v.inspect}.presence should return alternative" }
+    NOT.each   { |v| assert_equal v,           v.presence_or(alternative), "#{v.inspect}.presence should return self" }
+  end
 end